Move Arrow class vertex setting to Icon class
- so other Icons can use the Arrow class
