+++ src/Button.py
@@ -1,25 +1,5 @@
 """This file defines the Button class."""
-from enum import Enum
-from src.Icon import Icon
-
-
-class Arrow(Enum):
-    """This class defines Enums for the arrow polygon of a button."""
-
-    # Array of vectors for each arrow
-    FORWARD = [(20, -20), (0, -40), (-20, -20), (-10, -20), (-10, 40),
-               (10, 40), (10, -20)]
-
-    LEFT = [(-20, -40), (-40, -20), (-20, 0), (-20, -10), (10, -10),
-            (10, 20), (-30, 20), (-30, 40), (30, 40), (30, -30),
-            (-20, -30)]
-
-    RIGHT = [(20, -40), (40, -20), (20, 0), (20, -10), (-10, -10),
-             (-10, 20), (30, 20), (30, 40), (-30, 40), (-30, -30),
-             (20, -30)]
-
-    BACKWARD = [(20, 20), (0, 40), (-20, 20), (-10, 20), (-10, -40),
-                (10, -40), (10, 20)]
+from src.Icon import Icon, Arrow
 
 
 class Button(Icon):
@@ -39,20 +19,6 @@ class Button(Icon):
                          displayed)
 
         self.swapped = False  # Keeps track of wether a Button is swapped
-
-        if text == 'Forward':
-            self.vertices = Arrow.FORWARD.value
-        elif text == 'Turn Left':
-            self.vertices = Arrow.LEFT.value
-        elif text == 'Turn Right':
-            self.vertices = Arrow.RIGHT.value
-        elif text == 'Backward':
-            self.vertices = Arrow.BACKWARD.value
-
-        if self.vertices:
-            self.vertices = self._get_vertex_list(self.vertices,
-                                                  self.rect.centerx,
-                                                  self.rect.centery)
 
     def swap_colours(self):
         """Swap the background and text / polygon colour of the Button."""
